hrSystem/requirements_test.txt
behave==1.2.6
behave-django==1.4.0
selenium==4.1.0
pytest-bdd==6.0.1
factory-boy==3.2.1